server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
